chore: Move avm1 objects to their own module, we have plenty of them now
