+++ core/src/avm1/globals/movie_clip_loader.rs
@@ -2,9 +2,9 @@
 
 use crate::avm1::activation::Activation;
 use crate::avm1::error::Error;
+use crate::avm1::object::script_object::ScriptObject;
 use crate::avm1::object::TObject;
 use crate::avm1::property::Attribute;
-use crate::avm1::script_object::ScriptObject;
 use crate::avm1::{Object, UpdateContext, Value};
 use crate::backend::navigator::RequestOptions;
 use crate::display_object::{DisplayObject, TDisplayObject};
