+++ core/src/avm1/object/script_object.rs
