+++ core/src/avm1/object/shared_object.rs
@@ -1,8 +1,8 @@
 use crate::avm1::activation::Activation;
 use crate::avm1::error::Error;
 use crate::avm1::function::Executable;
+use crate::avm1::object::sound_object::SoundObject;
 use crate::avm1::property::Attribute;
-use crate::avm1::sound_object::SoundObject;
 use crate::avm1::{Object, ObjectPtr, ScriptObject, TObject, Value};
 use crate::context::UpdateContext;
 use crate::display_object::DisplayObject;
