+++ core/src/avm1/object/stage_object.rs
