+++ core/src/avm1/object/value_object.rs
