+++ core/src/avm1/object/xml_attributes_object.rs
