+++ core/src/avm1/object/xml_idmap_object.rs
