+++ core/src/avm1/object/xml_object.rs
